# backend/openapi/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    Application:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        userKey:
          type: string
        planName:
          type: string
        planSystemName:
          type: string
        accountEmail:
          type: string
        applicationId:
          type: string
        applicationKey:
          type: string
        redirectUrl:
          type: string
    ApplicationPlan:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        systemName:
          type: string
        state:
          type: string
        limits:
          type: array
          items:
            $ref: "#/components/schemas/PlanLimit"
...
    BackendUsage:
      type: object
      properties:
        backendName:
          type: string
        path:
          type: string
...
    MappingRule:
      type: object
      properties:
        httpMethod:
          type: string
        pattern:
          type: string
        metricRef:
          type: string
        delta:
          type: integer
          format: int32
...
    PlanLimit:
      type: object
      properties:
        metricName:
          type: string
        period:
          type: string
        value:
          type: integer
          format: int64
...
    ThreeScaleProduct:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        systemName:
          type: string
        serviceId:
          type: integer
          format: int64
        description:
          type: string
        deploymentOption:
          type: string
        mappingRules:
          type: array
          items:
            $ref: "#/components/schemas/MappingRule"
        backendUsages:
          type: array
          items:
            $ref: "#/components/schemas/BackendUsage"
        authentication:
          type: object
          additionalProperties: {}
        source:
          type: string
        backendNamespace:
          type: string
        backendServiceName:
          type: string
        sourceCluster:
          type: string
        applicationPlans:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationPlan"
        applications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
    ThreeScaleRefreshResult:
      description: Result of evicting discovery cache and reloading 3scale data
      type: object
      properties:
        productCount:
          type: integer
          format: int32
        backendCount:
          type: integer
          format: int32
        refreshedAt:
          type: string
    ThreeScaleSource:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        adminUrl:
          type: string
        accessToken:
          type: string
        enabled:
          type: boolean
    ThreeScaleSourceStatus:
      type: object
      description: Reachability status for a configured 3scale Admin API source
      properties:
        id:
          type: string
        label:
          type: string
        adminUrl:
          type: string
        configured:
          type: boolean
        enabled:
          type: boolean
        reachable:
          type: boolean
        error:
          type: string
    ThreeScaleSourceView:
      description: 3scale Admin API source (public view — credentials never exposed)
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        adminUrl:
          type: string
        enabled:
          type: boolean
        credentialConfigured:
          type: boolean
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      description: Authentication
paths:
...
  /api/threescale/products/{namespace}/{name}:
    get:
      operationId: getProduct
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreeScaleProduct"
      summary: Get Product
      tags:
      - Three Scale Resource
  /api/threescale/refresh:
    post:
      operationId: refreshDiscovery
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreeScaleRefreshResult"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Refresh Discovery
      tags:
      - Three Scale Resource
      security:
      - SecurityScheme:
        - operator
        - admin
  /api/threescale/sources:
    get:
      operationId: listSources
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreeScaleSourceView"
      summary: List Sources
      tags:
      - Three Scale Resource
    post:
      operationId: addSource
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThreeScaleSource"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreeScaleSourceView"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Add Source
      tags:
      - Three Scale Resource
      security:
      - SecurityScheme:
        - admin
  /api/threescale/sources/{id}:
    delete:
      operationId: removeSource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Remove Source
      tags:
      - Three Scale Resource
      security:
      - SecurityScheme:
        - admin
  /api/threescale/sources/{id}/status:
    get:
      operationId: getSourceStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreeScaleSourceStatus"
      summary: Get Source Status
      tags:
      - Three Scale Resource
...
info:
  title: ApiShift API
  version: 0.3.0
  description: 3scale to Connectivity Link migration console REST API
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value